Refactor: Flatten semantic modules and simplify patterns
- Merged `src/semantic/assembled.rs` into `src/semantic/assembler.rs` to flatten the module hierarchy.
- Moved iterator pattern detection logic (`detect_iterator_pattern`) from `src/semantic/patterns.rs` to `src/semantic/conversion.rs` to improve code cohesion (collocating `AssembledStatement` analysis).
- Refactored `try_parse_struct_instantiation` in `src/semantic/patterns.rs` to use `analyze_argument_expr`, eliminating duplicated argument parsing logic and enabling support for complex expressions in struct constructors.
- Updated `src/semantic/mod.rs` exports and internal imports accordingly.
- Updated `.jules/razor.md` with reduction metrics.
